Code cleanup
- remove unused code, parameters, comments and redundant checks and casts
- simplify for loops, Math.min/max calls, empty collection checks and lambda expressions
- adds some fixes for nullability, code visibility, JavaDocs

+++ main/src/main/java/cgeo/geocaching/utils/GeoHeightUtils.java
@@ -29,7 +29,7 @@ public class GeoHeightUtils {
                 countReadings++;
             }
         }
-        averageAltitude /= (double) altitudeReadings.length;
+        averageAltitude /= altitudeReadings.length;
         return averageAltitude == 0 || (onlyFullReadings && countReadings < 5) ? "" : Formatter.SEPARATOR + (countReadings < 5 ? "~ " : "") + Units.formatElevation((float) averageAltitude);
     }
